apps-project/machine-learning/requirements.txt
fastapi
uvicorn
qdrant-client
python-dotenv
numpy
requests
scikit-learn
kneed